move into spring-mvc-simple
