+++ spring-mvc-simple/src/main/java/com/baeldung/spring/servlets/ForwardedServlet.java
@@ -1,4 +1,4 @@
-package com.baeldung.servlets;
+package com.baeldung.spring.servlets;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -6,12 +6,15 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 
-@WebServlet("/welcome")
-public class UnwelcomeServlet extends HttpServlet {
+@WebServlet("/forwarded")
+public class ForwardedServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.sendRedirect(req.getContextPath() + "/redirected");
+        PrintWriter out = resp.getWriter();
+        out.write("In forwarded servlet page.");
+        out.write("\nWelcome:" + req.getParameter("name"));
     }
 
 }
