+++ spring-mvc-simple/src/main/java/com/baeldung/spring/servlets/RedirectedServlet.java
@@ -1,19 +1,20 @@
-package com.baeldung.servlets;
+package com.baeldung.spring.servlets;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
 
-@WebServlet("/hello")
-public class HelloServlet extends HttpServlet {
+@WebServlet("/redirected")
+public class RedirectedServlet extends HttpServlet {
 
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/forwarded");
-        dispatcher.forward(req, resp);
+        PrintWriter out = resp.getWriter();
+        out.write("In redirected servlet page.");
+        out.write("\nWelcome:" + req.getParameter("name"));
     }
 
 }
