+++ spring-mvc-simple/src/test/java/com/baeldung/spring/servlets/HelloServletTest.java
@@ -1,13 +1,12 @@
-package com.baeldung.servlets;
+package com.baeldung.spring.servlets;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
-
 import javax.servlet.ServletException;
 import java.io.IOException;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class HelloServletTest {
     @Test
@@ -16,9 +15,9 @@ public class HelloServletTest {
         request.addParameter("name", "Dennis");
         MockHttpServletResponse response = new MockHttpServletResponse();
         HelloServlet servlet = new HelloServlet();
-        
+
         servlet.doGet(request, response);
-        
+
         assertEquals("/forwarded", response.getForwardedUrl());
         assertEquals(200, response.getStatus());
     }
