+++ spring-mvc-simple/src/test/java/com/baeldung/spring/servlets/WelcomeServletTest.java
@@ -1,20 +1,19 @@
-package com.baeldung.servlets;
+package com.baeldung.spring.servlets;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
-
 import javax.servlet.ServletException;
 import java.io.IOException;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class UnwelcomeServletTest {
+public class WelcomeServletTest {
     @Test
     public void whenRequested_thenRedirectedToCorrectUrl() throws ServletException, IOException {
         MockHttpServletRequest request = new MockHttpServletRequest("GET", "/welcome");
         request.addParameter("name", "Dennis");
-        UnwelcomeServlet servlet = new UnwelcomeServlet();
+        WelcomeServlet servlet = new WelcomeServlet();
         MockHttpServletResponse response = new MockHttpServletResponse();
         
         servlet.doGet(request, response);
